Camp Type and Core Competence Added

+++ resources/views/admin/questions/edit.blade.php
@@ -39,16 +39,32 @@
                                 {!! Form::label('competence', 'Kompetenz:') !!}
                                 {!! Form::text('competence', null, ['class' => 'form-control', 'required']) !!}
                         </div>
+                        
+                        <div class="form-row">
+                            <div class="form-group col-md-6">
+                                {!! Form::label('competence_js1', 'Kernkompetenz J+S-Leiter 1:') !!}
+                                {{-- {!! Form::hidden('competence_js1', false) !!} --}}
+                                {!! Form::checkbox('competence_js1', '1', ['class' => 'form-control', 'required']) !!}
+                            </div>
+                        
+                            <div class="form-group col-md-6">
+                                {!! Form::label('competence_js2', 'Kernkompetenz J+S-Leiter 2:') !!}
+                                {{-- {!! Form::hidden('competence_js2', false) !!} --}}
+                                {!! Form::checkbox('competence_js2',  '1', ['class' => 'form-control', 'required']) !!}
+                            </div>
+                        </div>
                         <div class="form-group">
                             {!! Form::submit('Update Frage', ['class' => 'btn btn-primary'])!!}
                         </div>
                     {!! Form::close()!!}
 
-                    {!! Form::model($question, ['method' => 'DELETE', 'action'=>['AdminQuestionsController@destroy',$question->id]]) !!}
-                    <div class="form-group">
-                        {!! Form::submit('Frage löschen', ['class' => 'btn btn-danger'])!!}
-                    </div>
-                    {!! Form::close()!!}
+                    @if (Auth::user()->isAdmin())
+                        {!! Form::model($question, ['method' => 'DELETE', 'action'=>['AdminQuestionsController@destroy',$question->id]]) !!}
+                        <div class="form-group">
+                            {!! Form::submit('Frage löschen', ['class' => 'btn btn-danger'])!!}
+                        </div>
+                        {!! Form::close()!!}
+                    @endif
                  </div>
             </div>
         </div>
